feat: day 06 solution in Wren
This commit adds my solution to both parts of the day 06 Advent of Code
challenge, implemented in Wren.

I've updated my main directory `README.md` to include the new language,
and a basic `day06/README.md` to help run the script.

Since Wren allows the use of classes, and this challenge was more
complex than previous challenges, I've taken the opportunity to create
some `.wren` classes to break the code up a bit.